requirements.txt
dotenv
langchain_community
sentence-transformers
chromadb
flask
langchain-huggingface
langchain-chroma
langchain-google-genai